requirements.txt
Django~=5.0.6
python-dotenv~=1.0.1
psycopg2~=2.9.9
djangorestframework~=3.15.2
drf-yasg~=1.21.7
setuptools~=71.0.4